requirements.txt
numpy
tensorflow-cpu
tensorflow_hub
streamlit
PIL